# packages/assistants-backend/src/schema/openapi.yaml
openapi: 3.1.0
info:
  title: assistants
  version: '1'
  description: >-
    The Backstage backend plugin that powers AI Assistants — multi-assistant
    chat with model selection, streaming, and per-assistant access policies.
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact: {}
servers:
  - url: /
components:
  examples: {}
  headers: {}
  parameters: {}
  requestBodies: {}
  responses:
    ErrorResponse:
      description: An error response from the backend.
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ModelOption:
      type: object
      description: A single selectable model in the global model pool.
      required:
        - id
        - provider
        - model
      properties:
        id:
          type: string
          description: Full `provider:model` id, e.g. "anthropic:claude-sonnet-4".
        provider:
          type: string
          description: Provider id, e.g. "anthropic".
        model:
          type: string
          description: Model name, e.g. "claude-sonnet-4".
    UiOptions:
      type: object
      description: >-
        Resolved, browser-safe UI options for an assistant's chat surface
        (deep-merge of global + per-profile `ui`).
      properties:
        composer:
          type: object
          description: Composer (message input) options.
          properties:
            placeholder:
              type: string
              description: Placeholder text shown in the empty message input.
        suggestions:
          type: array
          description: Starter prompt chips offered on an empty conversation.
          items:
            type: object
            required:
              - title
              - prompt
            properties:
              title:
                type: string
                description: Short label shown on the suggestion chip.
              prompt:
                type: string
                description: The text submitted when the chip is clicked.
    AssistantSummary:
      type: object
      description: >-
        Browser-safe projection of an assistant the caller may access. Never
        includes the prompt or access policy.
      required:
        - id
        - title
      properties:
        id:
          type: string
          description: Identifier for the configured assistant.
        title:
          type: string
          description: Human-readable assistant title.
        description:
          type: string
          description: Optional human-readable assistant description.
        models:
          type: array
          description: >-
            The `provider:model` ids this assistant may use (its allowlist).
            Omitted when the assistant allows the full pool.
          items:
            type: string
        defaultModel:
          type: string
          description: This assistant's default `provider:model` selection.
        ui:
          $ref: '#/components/schemas/UiOptions'
    StatusResponse:
      type: object
      description: Payload returned by `GET /status`.
      required:
        - assistants
        - models
        - defaultModel
      properties:
        assistants:
          type: array
          description: Assistants filtered to those the caller may access.
          items:
            $ref: '#/components/schemas/AssistantSummary'
        models:
          type: array
          description: The global model pool.
          items:
            $ref: '#/components/schemas/ModelOption'
        defaultModel:
          type: string
          description: The default `provider:model` selection.
    ChatRequest:
      type: object
      description: >-
        Request body shared by `POST /chat` and `POST /title`. Tools/actions are
        never accepted from the client — they are resolved server-side per turn.
      required:
        - assistantId
        - modelId
        - messages
      properties:
        assistantId:
          type: string
          minLength: 1
          description: Id of the assistant to converse with.
        modelId:
          type: string
          minLength: 1
          description: The `provider:model` id to use for this turn.
        messages:
          type: array
          minItems: 1
          description: >-
            The conversation as a non-empty array of UI messages. The internal
            shape of each message is intentionally not over-constrained here.
          items:
            type: object
    TitleResponse:
      type: object
      description: Response returned by `POST /title`.
      required:
        - title
      properties:
        title:
          type: string
          description: The generated (or fallback) conversation title.
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            name:
              type: string
            message:
              type: string
          required:
            - name
            - message
        request:
          type: object
          properties:
            method:
              type: string
            url:
              type: string
          required:
            - method
            - url
        response:
          type: object
          properties:
            statusCode:
              type: number
          required:
            - statusCode
      required:
        - error
        - request
        - response
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /status:
    get:
      operationId: GetStatus
      description: >-
        The single page-load endpoint. Returns the assistants the caller may
        access (summary only — never the prompt or access policy), the full
        model pool, and the default model.
      responses:
        '200':
          description: The browser-safe plugin status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - {}
        - JWT: []
  /title:
    post:
      operationId: PostTitle
      description: >-
        Generate a short conversation title from the opening messages.
        Best-effort: generation failures fall back to a default title with 200.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: The generated (or fallback) title.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - {}
        - JWT: []
  /chat:
    post:
      operationId: PostChat
      description: >-
        Stream a chat turn for the given assistant and model. The request body
        is validated against ChatRequest; the RESPONSE is a UI message stream
        piped via `pipeUIMessageStreamToResponse` and is intentionally NOT
        response-validated — it is modeled loosely as `text/event-stream`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: A streamed UI message response (Server-Sent Events).
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - {}
        - JWT: []