# packages/assistants-backend/src/schema/openapi/generated/models/AssistantSummary.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { UiOptions } from "../models/UiOptions.model";

/**
 * Browser-safe projection of an assistant the caller may access. Never includes the prompt or access policy.
 * @public
 */
export interface AssistantSummary {
  /**
   * Identifier for the configured assistant.
   */
  id: string;
  /**
   * Human-readable assistant title.
   */
  title: string;
  /**
   * Optional human-readable assistant description.
   */
  description?: string;
  /**
   * The `provider:model` ids this assistant may use (its allowlist). Omitted when the assistant allows the full pool.
   */
  models?: Array<string>;
  /**
   * This assistant's default `provider:model` selection.
   */
  defaultModel?: string;
  ui?: UiOptions;
}