# packages/assistants-backend/src/schema/openapi/generated/models/ChatRequest.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * Request body shared by `POST /chat` and `POST /title`. Tools/actions are never accepted from the client — they are resolved server-side per turn.
 * @public
 */
export interface ChatRequest {
  /**
   * Id of the assistant to converse with.
   */
  assistantId: string;
  /**
   * The `provider:model` id to use for this turn.
   */
  modelId: string;
  /**
   * The conversation as a non-empty array of UI messages. The internal shape of each message is intentionally not over-constrained here.
   */
  messages: Array<any>;
}