# packages/assistants-backend/src/schema/openapi/generated/models/ModelOption.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * A single selectable model in the global model pool.
 * @public
 */
export interface ModelOption {
  /**
   * Full `provider:model` id, e.g. \"anthropic:claude-sonnet-4\".
   */
  id: string;
  /**
   * Provider id, e.g. \"anthropic\".
   */
  provider: string;
  /**
   * Model name, e.g. \"claude-sonnet-4\".
   */
  model: string;
}