# packages/assistants-backend/src/schema/openapi/generated/models/StatusResponse.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { AssistantSummary } from "../models/AssistantSummary.model";
import { ModelOption } from "../models/ModelOption.model";

/**
 * Payload returned by `GET /status`.
 * @public
 */
export interface StatusResponse {
  /**
   * Assistants filtered to those the caller may access.
   */
  assistants: Array<AssistantSummary>;
  /**
   * The global model pool.
   */
  models: Array<ModelOption>;
  /**
   * The default `provider:model` selection.
   */
  defaultModel: string;
}