# packages/assistants-backend/src/schema/openapi/generated/models/TitleResponse.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * Response returned by `POST /title`.
 * @public
 */
export interface TitleResponse {
  /**
   * The generated (or fallback) conversation title.
   */
  title: string;
}