# packages/assistants-backend/src/schema/openapi/generated/models/UiOptions.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************

/**
 * Resolved, browser-safe UI options for an assistant's chat surface (deep-merge of global + per-profile `ui`).
 * @public
 */
export interface UiOptions {
  /**
   * Composer (message input) options.
   */
  composer?: {
    /**
     * Placeholder text shown in the empty message input.
     */
    placeholder?: string;
  };
  /**
   * Starter prompt chips offered on an empty conversation.
   */
  suggestions?: Array<{
    /**
     * Short label shown on the suggestion chip.
     */
    title: string;
    /**
     * The text submitted when the chip is clicked.
     */
    prompt: string;
  }>;
}